+++ src/actions/generateActions.js
@@ -7,7 +7,7 @@ import { runValidations } from '../utils/runValidations.js';
 import { getMainClassDirectory, getRootPackage } from '../utils/javaPackageResolver.js';
 import { logger } from '../utils/logger.js';
 import { renderTemplate } from '../utils/templateCompiler.js';
-import { structure } from '../config/architectures/feature_based.js';
+import { structure } from '../config/architectures/featureBased.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -16,18 +16,18 @@ export function generateFeature(featureName) {
     const javaPath = getMainClassDirectory();
     const packageName = getRootPackage();
 
-    const input_context = { featureName, javaPath, packageName };
+    const inputContext = { featureName, javaPath, packageName };
 
     runValidations(validations, featureName);
 
-    generateFeatureStructure(input_context)
+    generateFeatureStructure(inputContext)
     logger.success(`\n` + 'Feature created successfully');
 }
 
 function generateFeatureStructure({ featureName, javaPath, packageName }) {
     const capitalizedName = featureName.charAt(0).toUpperCase() + featureName.slice(1).toLowerCase();
     const templatesBasePath = path.join(__dirname, '..', 'templates', 'feature');
-    const template_context = { featureName, packageName, capitalizedName };
+    const templateContext = { featureName, packageName, capitalizedName };
 
     for (const [folder, file] of Object.entries(structure)) {
         const folderPath = path.join(javaPath, file.folder(featureName), folder);
@@ -37,7 +37,7 @@ function generateFeatureStructure({ featureName, javaPath, packageName }) {
         logger.info(`Creating ${folder}...`);
         fs.mkdirSync(folderPath, { recursive: true });
 
-        const content = renderTemplate(fullPath, template_context);
+        const content = renderTemplate(fullPath, templateContext);
         fs.writeFileSync(path.join(folderPath, fileName), content);
     }
 }